Fixed code smell: This class has 6 parents which is greater than 5 authorized.
